+++ Showcase/Showcase/Pages/Tokens/Typography/TypographyTokenPage.swift
@@ -43,7 +43,7 @@ struct TypographyTokenPage: View {
 
         VStack(alignment: .leading, spacing: theme.spaceFixedNone) {
             typgraphyIllustration(from: namedTypography)
-                .foregroundStyle((theme.colorContentMuted.color(for: colorScheme)))
+                .foregroundStyle(theme.colorContentMuted.color(for: colorScheme))
 
             Group {
                 Text("family (\(theme.customFontFamily ?? "system")), ")
@@ -54,7 +54,7 @@ struct TypographyTokenPage: View {
             }
             .typeBodyDefaultMedium(theme)
             .fixedSize(horizontal: false, vertical: true)
-            .foregroundStyle((theme.colorContentDefault.color(for: colorScheme)))
+            .foregroundStyle(theme.colorContentDefault.color(for: colorScheme))
         }
         .frame(maxWidth: .infinity, alignment: .leading)
         .padding(.vertical, theme.spaceFixedShorter)