+++ Showcase/Showcase/Pages/Utils/ShowcaseTokenIllustration.swift
@@ -22,10 +22,9 @@ struct ShowcaseTokenIllustration<TokenIllustration>: View where TokenIllustratio
 
     // MARK: Stored properties
 
-    @ViewBuilder
-    let tokenIllustration: () -> TokenIllustration
     let tokenName: String
     let tokenValue: String?
+    @ViewBuilder let tokenIllustration: () -> TokenIllustration
 
     // MARK: Initializer
 
@@ -45,11 +44,11 @@ struct ShowcaseTokenIllustration<TokenIllustration>: View where TokenIllustratio
             VStack(alignment: .leading) {
                 Text(tokenName)
                     .typeBodyStrongLarge(theme)
-                    .foregroundStyle((theme.colorContentDefault.color(for: colorScheme)))
+                    .foregroundStyle(theme.colorContentDefault.color(for: colorScheme))
                 if let tokenValue {
                     Text(tokenValue)
                         .typeBodyDefaultMedium(theme)
-                        .foregroundStyle((theme.colorContentMuted.color(for: colorScheme)))
+                        .foregroundStyle(theme.colorContentMuted.color(for: colorScheme))
                 }
             }
             .frame(maxWidth: .infinity, alignment: .leading)